requirements.txt
aiogram==3.7.0
python-dotenv==1.0.1
PyYAML==6.0.2
APScheduler==3.10.4
pillow==10.4.0
openai==1.40.6
matplotlib==3.9.2
psycopg[binary]==3.2.1
pandas==2.2.2
requests==2.32.3
reportlab==4.2.2
